+++ Runtime/Proxy/Common/SdkInfo.cs
@@ -8,6 +8,6 @@ namespace Playserv.Proxy.Common
         /// <summary>
         /// Current SDK version.
         /// </summary>
-        public const string Version = "0.6.6";
+        public const string Version = "0.6.7";
     }
 }